working snap on CustomSwitch
