+++ MapifyEditor/Trackage/CustomSwitch.cs
@@ -1,4 +1,7 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using UnityEditor;
 using UnityEngine;
 
 namespace Mapify.Editor
@@ -20,6 +23,18 @@ namespace Mapify.Editor
                 throw new IndexOutOfRangeException($"Branch index {branchIndex} is out of range. Switch has {Tracks.Length} tracks.");
             }
             return Tracks[branchIndex].Curve.Last();
+        }
+
+        public List<BezierPoint> GetOutPoints()
+        {
+            return Tracks.Select(track => track.Curve.Last()).ToList();
+        }
+
+        public override List<BezierPoint> GetPoints()
+        {
+            var points = new List<BezierPoint>{ GetJointPoint() };
+            points.AddRange(GetOutPoints());
+            return points;
         }
 
         [Tooltip("Which way the switch should be flipped by default")]
