+++ MapifyEditor/Trackage/SwitchBase.cs
@@ -6,6 +6,7 @@ using UnityEngine;
 
 namespace Mapify.Editor
 {
+    [ExecuteInEditMode] //this is necessary for snapping to work
     public abstract class SwitchBase: MonoBehaviour
     {
         public virtual Track[] GetTracks()
@@ -16,49 +17,134 @@ namespace Mapify.Editor
 
         public abstract BezierPoint GetJointPoint();
 
-        // TODO why is this unused
-//         protected void Snap()
-//         {
-// #if UNITY_EDITOR
-//             BezierPoint[] bezierPoints = FindObjectsOfType<BezierPoint>();
-//             GameObject[] selectedObjects = Selection.gameObjects;
-//             bool isSelected = selectedObjects.Contains(gameObject);
-//
-//             TrySnap(bezierPoints, isSelected, transform);
-//             foreach (var track in Tracks)
-//             {
-//                 TrySnap(bezierPoints, isSelected, track.Curve.Last().transform);
-//             }
-// #endif
-//         }
-//
-//         private void TrySnap(IEnumerable<BezierPoint> points, bool move, Transform reference)
-//         {
-//             var referencePosition = reference.position;
-//             var closestPos = Vector3.zero;
-//             var closestDist = float.MaxValue;
-//
-//             foreach (var otherPoint in points)
-//             {
-//                 if (otherPoint.Curve().GetComponentInParent<SwitchBase>() == this) continue;
-//
-//                 var otherPosition = otherPoint.transform.position;
-//                 var dist = Mathf.Abs(Vector3.Distance(otherPosition, referencePosition));
-//
-//                 if (dist > Track.SNAP_RANGE || dist >= closestDist) continue;
-//
-//                 var aTrack = otherPoint.GetComponentInParent<Track>();
-//                 if (aTrack.IsSwitch) continue;
-//
-//                 closestPos = otherPosition;
-//                 closestDist = dist;
-//             }
-//
-//             if (closestDist >= float.MaxValue) return;
-//
-//             if (move) {
-//                 transform.position = closestPos + (transform.position - reference.position);
-//             }
-//         }
+        public abstract List<BezierPoint> GetPoints();
+
+#if UNITY_EDITOR
+        private bool snapShouldUpdate = true;
+
+        private Vector3[] previousPositionsPoints;
+        private SnappedTrack[] snappedTracks;
+        private bool started = false;
+
+        private void Start()
+        {
+            var pointsCount = GetTracks().Count()+1;
+            previousPositionsPoints = new Vector3[pointsCount];
+            snappedTracks = new SnappedTrack[pointsCount];
+            started = true;
+        }
+
+        private void OnEnable()
+        {
+            snapShouldUpdate = true;
+        }
+
+        private void OnDisable()
+        {
+            UnsnapConnectedTracks();
+        }
+
+        private void OnDestroy()
+        {
+            UnsnapConnectedTracks();
+        }
+
+        private void OnDrawGizmos()
+        {
+            if(!started) return;
+            if (transform.DistToSceneCamera() >= Track.SNAP_UPDATE_RANGE_SQR)
+            {
+                return;
+            }
+
+            CheckSwitchMoved();
+
+            if (snapShouldUpdate)
+            {
+                Snap();
+                snapShouldUpdate = false;
+            }
+        }
+
+        private void CheckSwitchMoved()
+        {
+            var positionPoints = GetPoints().Select(point => point.position).ToArray();
+
+            for (int index = 0; index < positionPoints.Length; index++)
+            {
+                if (positionPoints[index] == previousPositionsPoints[index]) continue;
+
+                snapShouldUpdate = true;
+                previousPositionsPoints[index] = positionPoints[index];
+            }
+        }
+
+        private void UnsnapConnectedTracks()
+        {
+            foreach (var snapped in snappedTracks)
+            {
+                snapped?.UnSnapped();
+            }
+        }
+
+        public void Snap()
+        {
+            var bezierPoints = FindObjectsOfType<BezierPoint>();
+            bool isSelected = Selection.gameObjects.Contains(gameObject);
+
+            var points = GetPoints();
+            for (var pointIndex = 0; pointIndex < points.Count; pointIndex++)
+            {
+                TrySnap(bezierPoints, isSelected, points[pointIndex], pointIndex);
+            }
+        }
+
+        private void TrySnap(IEnumerable<BezierPoint> points, bool move, BezierPoint snapPoint, int snapPointIndex)
+        {
+            var reference = snapPoint.transform;
+
+            var position = reference.position;
+            var closestPosition = Vector3.zero;
+            var closestDistance = float.MaxValue;
+
+            foreach (BezierPoint otherSnapPoint in points)
+            {
+                //don't connect to itself
+                if (otherSnapPoint.Curve().GetComponentInParent<Switch>() == this) continue;
+
+                Vector3 otherPosition = otherSnapPoint.transform.position;
+                float distance = Mathf.Abs(Vector3.Distance(otherPosition, position));
+
+                // too far away
+                if (distance > Track.SNAP_RANGE || distance >= closestDistance) continue;
+
+                var otherTrack = otherSnapPoint.GetComponentInParent<Track>();
+
+                // don't snap a switch to another switch
+                if (otherTrack.IsSwitch) continue;
+
+                closestPosition = otherPosition;
+                closestDistance = distance;
+
+                otherTrack.Snapped(otherSnapPoint);
+
+                //remember what track we snapped to
+                snappedTracks[snapPointIndex] = new SnappedTrack(otherTrack, otherSnapPoint);
+            }
+
+            // No snap target found
+            if (closestDistance >= float.MaxValue)
+            {
+                snappedTracks[snapPointIndex]?.UnSnapped();
+                snappedTracks[snapPointIndex] = null;
+                return;
+            }
+
+            if (move)
+            {
+                transform.position = closestPosition + (transform.position - reference.position);
+            }
+        }
+#endif
     }
 }
